+++ packages/popup-sheet/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @alfalab/core-components-popup-sheet
 
+## 2.0.7
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-navigation-bar-private@1.1.2
+
 ## 2.0.6
 
 ### Patch Changes